@@ -1,32 +1,58 @@
 {
   "name": "whatwg-url",
-  "version": "5.0.0",
+  "version": "11.0.0",
   "description": "An implementation of the WHATWG URL Standard's URL API and parsing machinery",
-  "main": "lib/public-api.js",
+  "main": "index.js",
   "files": [
-    "lib/"
+    "index.js",
+    "webidl2js-wrapper.js",
+    "lib/*.js"
   ],
   "author": "Sebastian Mayr <github@smayr.name>",
   "license": "MIT",
   "repository": "jsdom/whatwg-url",
   "dependencies": {
-    "tr46": "~0.0.3",
-    "webidl-conversions": "^3.0.0"
+    "tr46": "^3.0.0",
+    "webidl-conversions": "^7.0.0"
   },
   "devDependencies": {
-    "eslint": "^2.6.0",
-    "istanbul": "~0.4.3",
-    "mocha": "^2.2.4",
-    "recast": "~0.10.29",
-    "request": "^2.55.0",
-    "webidl2js": "^3.0.2"
+    "@domenic/eslint-config": "^1.4.0",
+    "benchmark": "^2.1.4",
+    "browserify": "^17.0.0",
+    "domexception": "^4.0.0",
+    "eslint": "^7.32.0",
+    "got": "^11.8.2",
+    "jest": "^27.2.4",
+    "webidl2js": "^17.0.0"
+  },
+  "engines": {
+    "node": ">=12"
   },
   "scripts": {
-    "build": "node scripts/transform.js && node scripts/convert-idl.js",
-    "coverage": "istanbul cover node_modules/mocha/bin/_mocha",
+    "coverage": "jest --coverage",
     "lint": "eslint .",
-    "prepublish": "npm run build",
-    "pretest": "node scripts/get-latest-platform-tests.js && npm run build",
-    "test": "mocha"
+    "prepare": "node scripts/transform.js",
+    "pretest": "node scripts/get-latest-platform-tests.js && node scripts/transform.js",
+    "build-live-viewer": "browserify index.js --standalone whatwgURL > live-viewer/whatwg-url.js",
+    "test": "jest"
+  },
+  "jest": {
+    "collectCoverageFrom": [
+      "lib/**/*.js",
+      "!lib/utils.js"
+    ],
+    "coverageDirectory": "coverage",
+    "coverageReporters": [
+      "lcov",
+      "text-summary"
+    ],
+    "testEnvironment": "node",
+    "testMatch": [
+      "<rootDir>/test/**/*.js"
+    ],
+    "testPathIgnorePatterns": [
+      "^<rootDir>/test/testharness.js$",
+      "^<rootDir>/test/web-platform-tests/"
+    ]
   }
 }
